requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
pillow>=10.3.0
python-multipart>=0.0.9
greenlet>=3.0.3
